# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A high-performance Git version control system API with memory-first architecture",
        "title": "GOVC API",
        "termsOfService": "https://github.com/caiatech/govc/blob/main/LICENSE",
        "contact": {
            "name": "GOVC API Support",
            "url": "https://github.com/caiatech/govc/issues",
            "email": "support@caiatech.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/caiatech/govc/blob/main/LICENSE"
        },
        "version": "1.0.0-dev"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/repositories": {
            "get": {
                "description": "Retrieves a list of all repositories with their metadata",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Repository Management"
                ],
                "summary": "List all repositories",
                "responses": {
                    "200": {
                        "description": "List of repositories",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "count": {
                                    "type": "integer"
                                },
                                "repositories": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/api.RepoResponse"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new Git repository with the specified configuration",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Repository Management"
                ],
                "summary": "Create a new repository",
                "parameters": [
                    {
                        "description": "Repository creation request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.CreateRepoRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Repository created successfully",
                        "schema": {
                            "$ref": "#/definitions/api.RepoResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request data",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Repository already exists",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Maximum repositories reached",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/repositories/{repo_id}": {
            "get": {
                "description": "Retrieves detailed information about a specific repository",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Repository Management"
                ],
                "summary": "Get repository information",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Repository ID",
                        "name": "repo_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Repository information",
                        "schema": {
                            "$ref": "#/definitions/api.RepoResponse"
                        }
                    },
                    "404": {
                        "description": "Repository not found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Permanently deletes a repository and all its data",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Repository Management"
                ],
                "summary": "Delete a repository",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Repository ID",
                        "name": "repo_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Repository deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/api.SuccessResponse"
                        }
                    },
                    "404": {
                        "description": "Repository not found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/v2/repositories": {
            "get": {
                "description": "Retrieves a list of all repositories using V2 clean architecture with enhanced metadata",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "V2 Architecture"
                ],
                "summary": "List all repositories (V2 Clean Architecture)",
                "responses": {
                    "200": {
                        "description": "List of repositories",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "count": {
                                    "type": "integer"
                                },
                                "repositories": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/api.RepoResponse"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new Git repository using the V2 clean architecture with enhanced performance and modularity",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "V2 Architecture"
                ],
                "summary": "Create a new repository (V2 Clean Architecture)",
                "parameters": [
                    {
                        "description": "Repository creation request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.CreateRepoRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Repository created successfully",
                        "schema": {
                            "$ref": "#/definitions/api.RepoResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request data",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Repository already exists",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Maximum repositories reached",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v2/repositories/{repo_id}": {
            "get": {
                "description": "Retrieves detailed information about a specific repository using V2 clean architecture",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "V2 Architecture"
                ],
                "summary": "Get repository information (V2 Clean Architecture)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Repository ID",
                        "name": "repo_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Repository information",
                        "schema": {
                            "$ref": "#/definitions/api.RepoResponse"
                        }
                    },
                    "404": {
                        "description": "Repository not found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Permanently deletes a repository using V2 clean architecture with proper cleanup",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "V2 Architecture"
                ],
                "summary": "Delete a repository (V2 Clean Architecture)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Repository ID",
                        "name": "repo_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Repository deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/api.SuccessResponse"
                        }
                    },
                    "404": {
                        "description": "Repository not found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.CreateRepoRequest": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "type": "string",
                    "example": "my-repo"
                },
                "memory_only": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "api.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "REPO_NOT_FOUND"
                },
                "details": {},
                "error": {
                    "type": "string",
                    "example": "Repository not found"
                }
            }
        },
...
        "api.RepoResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2023-01-01T00:00:00Z"
                },
                "current_branch": {
                    "type": "string",
                    "example": "main"
                },
                "id": {
                    "type": "string",
                    "example": "my-repo"
                },
                "path": {
                    "type": "string",
                    "example": "/repos/my-repo"
                }
            }
        },
...
        "api.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "Repository deleted successfully"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Enter the token with the `Bearer: ` prefix, e.g. \"Bearer abcde12345\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "ApiKeyQuery": {
            "description": "API key for accessing the API",
            "type": "apiKey",
            "name": "api_key",
            "in": "query"
        },
        "BasicAuth": {
            "type": "basic"
        }
    },
...
}